chore(gormx): Adjust files location
